add mySql changes
